update packages draft code for diceKeys shamir format
