+++ web/package.json
@@ -34,11 +34,11 @@
     "@dicekeys/read-dicekey-js": "^0.0.39",
     "@dicekeys/seeded-crypto-js": "^0.3.0",
     "mobx": "^6.3.4",
-    "mobx-react": "^7.2.1",
+    "mobx-react": "^9.0.1",
     "qrcode": "^1.5.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "styled-components": "^5.3.1"
+    "styled-components": "^6.1.0"
   },
   "browserslist": [
     "electron >= 15.0.0",
@@ -49,26 +49,25 @@
     "@types/core-js": "^2.5.3",
     "@types/emscripten": "^1.39.4",
     "@types/jest": "^29.0.3",
-    "@types/jsdom": "^20.0.0",
-    "@types/node": "^18.0.3",
+    "@types/jsdom": "^21.1.4",
+    "@types/node": "^20.8.10",
     "@types/qrcode": "^1.5.0",
-    "@types/react": "^18.0.8",
+    "@types/react": "^18.2.36",
     "@types/react-dom": "^18.0.0",
-    "@types/styled-components": "^5.1.15",
     "@types/w3c-image-capture": "^1.0.2",
     "@types/w3c-web-hid": "^1.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
-    "@vitejs/plugin-react": "^2.1.0",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react": "^4.1.1",
     "jest": "^29.0.3",
     "jest-environment-jsdom": "^29.0.3",
     "jest-raw-loader": "^1.0.1",
-    "jsdom": "^20.0.0",
-    "rimraf": "^3.0.2",
+    "jsdom": "^22.1.0",
+    "rimraf": "^5.0.5",
     "ts-jest": "^29.0.1",
     "ts-node": "^10.8.1",
-    "typescript": "^4.7.3",
-    "vite": "^3.1.3"
+    "typescript": "^5.2.2",
+    "vite": "^4.5.0"
   },
   "volta": {
     "node": "18.1.0",
