+++ web/src/dicekeys/DiceKey/Face.ts
@@ -1,0 +1,20 @@
+import type {
+  FaceIdentifiers,
+  FaceOrientationLetterTrbl,
+} from "@dicekeys/read-dicekey-js";
+export {
+  FaceLetter, FaceLetters, InvalidFaceLetterException,
+  FaceDigit, InvalidFaceDigitException,
+  FaceOrientationLetterTrblOrUnknown,
+  InvalidFaceOrientationLettersTrblOrUnknownException,
+  FaceOrientationLettersTrbl,
+  FaceDigits,
+  FaceOrientationLetterTrbl
+} from "@dicekeys/read-dicekey-js";
+export type {
+  FaceIdentifiers
+} from "@dicekeys/read-dicekey-js";
+
+export type Face = FaceIdentifiers & {
+  orientationAsLowercaseLetterTrbl: FaceOrientationLetterTrbl
+}
