+++ web/src/dicekeys/DiceKey/InvalidDiceKeyException.ts
@@ -1,0 +1,1 @@
+export class InvalidDiceKeyException extends Error {}
