+++ web/src/utilities/InferComponentProps.ts
@@ -1,21 +1,25 @@
 // Import types only.
 // From https://stackoverflow.com/questions/64054467/infer-prop-interface-from-styled-component
 
-import { ComponentPropsWithoutRef, JSXElementConstructor } from "react"
-import {
-  AnyStyledComponent,
-  StyledComponentInnerComponent,
-  StyledComponentInnerOtherProps,
-  StyledComponentInnerAttrs
-} from "styled-components"
+// import { ComponentPropsWithoutRef, JSXElementConstructor } from "react"
+// import {
+//   // AnyStyledComponent,
+//   // StyledComponentInnerComponent,
+//   // StyledComponentInnerOtherProps,
+//   // StyledComponentInnerAttrs
+//   StyledComponentPropsWithRef
+// } from "styled-components"
 
-export type InferComponentProps<T> =
-  T extends AnyStyledComponent ?
-    ( 
-        ComponentPropsWithoutRef<StyledComponentInnerComponent<T>> &
-        Omit<StyledComponentInnerOtherProps<T>, StyledComponentInnerAttrs<T>>
-    )
-    :
-    T extends keyof JSX.IntrinsicElements | JSXElementConstructor<unknown> ?
-      ComponentPropsWithoutRef<T> :
-      never;
+// import StyledComponentPropsWithRef from "styled-components"
+
+// export type InferComponentProps<T> = StyledComponentPropsWithRef<T>;
+
+  // T extends AnyStyledComponent ?
+  //   ( 
+  //       ComponentPropsWithoutRef<StyledComponentInnerComponent<T>> &
+  //       Omit<StyledComponentInnerOtherProps<T>, StyledComponentInnerAttrs<T>>
+  //   )
+  //   :
+  //   T extends keyof JSX.IntrinsicElements | JSXElementConstructor<unknown> ?
+  //     ComponentPropsWithoutRef<T> :
+  //     never;
