+++ web/src/views/DiceKeySelectorView.tsx
@@ -4,7 +4,8 @@ import React from "react";
 import styled, { css } from "styled-components";
 import { DiceKeyView } from "./SVG/DiceKeyView";
 import { cssCalcTyped, cssExprWithoutCalc } from "../utilities";
-import { DiceKeyInHumanReadableForm, DiceKeyWithKeyId, DiceKeyWithoutKeyId, facesFromPublicKeyDescriptor } from "../dicekeys/DiceKey";
+import { DiceKeyWithKeyId, DiceKeyWithoutKeyId, facesFromPublicKeyDescriptor } from "../dicekeys/DiceKey";
+import { DiceKeyInHumanReadableForm } from "../dicekeys/DiceKey";
 import { DiceKeyMemoryStore, PublicDiceKeyDescriptorWithSavedOnDevice } from "../state";
 import LoadDiceKeyImage from "../images/Scanning a DiceKey.svg";
 import { ToggleState } from "../state";
